+++ packages/gamut-labs/src/experimental/GlobalHeader/GlobalHeaderVariants.tsx
@@ -20,6 +20,8 @@ import {
   resourcesDropdown,
   search,
   signUp,
+  tryProForFree,
+  unpausePro,
   upgradeToPro,
 } from './GlobalHeaderItems';
 import { User } from './types';
@@ -133,7 +135,9 @@ export const freeHeaderItems = (
   renderSearch && rightItems.push(search(renderSearch));
   renderNotifications && rightItems.push(notifications(renderNotifications));
   rightItems.push(freeProfile(user));
-  rightItems.push(upgradeToPro);
+  rightItems.push(
+    user.showProUpgrade ? upgradeToPro : tryProForFree(user.proTrialCheckoutUrl)
+  );
 
   return {
     left: leftItems,
@@ -158,8 +162,11 @@ export const freeMobileHeaderItems = (
     pricingDropdown,
     forBusiness,
     freeProfile(user, true),
-    upgradeToPro,
   ];
+
+  mainMenuItems.push(
+    user.showProUpgrade ? upgradeToPro : tryProForFree(user.proTrialCheckoutUrl)
+  );
 
   return {
     left: leftItems,
@@ -185,6 +192,7 @@ export const proHeaderItems = (
   renderSearch && rightItems.push(search(renderSearch));
   renderNotifications && rightItems.push(notifications(renderNotifications));
   rightItems.push(proProfile(user));
+  user.isPaused && rightItems.push(unpausePro);
 
   return {
     left: leftItems,
@@ -208,6 +216,8 @@ export const proMobileHeaderItems = (
     communityDropdown,
     proProfile(user, true),
   ];
+
+  user.isPaused && mainMenuItems.push(unpausePro);
 
   return {
     left: leftItems,