+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.2.13](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.12...@codecademy/gamut-kit@0.2.13) (2021-03-10)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.12](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.11...@codecademy/gamut-kit@0.2.12) (2021-03-10)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.11](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.10...@codecademy/gamut-kit@0.2.11) (2021-03-10)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.10](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.9...@codecademy/gamut-kit@0.2.10) (2021-03-09)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.9](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.8...@codecademy/gamut-kit@0.2.9) (2021-03-09)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.8](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.7...@codecademy/gamut-kit@0.2.8) (2021-03-09)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.7](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.6...@codecademy/gamut-kit@0.2.7) (2021-03-09)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.6](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.5...@codecademy/gamut-kit@0.2.6) (2021-03-09)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.5](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.4...@codecademy/gamut-kit@0.2.5) (2021-03-08)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.4](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.3...@codecademy/gamut-kit@0.2.4) (2021-03-08)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
 ### [0.2.3](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.2...@codecademy/gamut-kit@0.2.3) (2021-03-05)
 
 **Note:** Version bump only for package @codecademy/gamut-kit